[PAPERCRAFT] Remove the need for a ComponentSystem class
The concept of the abstract component system is removed because I felt
that adding a messaging system for component interactions was too much
overhead. It would make debugging much more difficult and confusing the
setup as well. I am keeping the generic Component base class to store a
reference to shared component variables